修复yiban_clock配置文件生成规则

+++ yiban_clock/docker-entrypoint.sh
@@ -23,14 +23,15 @@ if [ ! $MYSQL_DATABASE ];then
   exit 100
 fi
 
-mkdir -p /app/config
-cp /config.json /app/config/config.json
-sed -i 's/$MYSQL_USER/'$MYSQL_USER'/g' /app/config/config.json
-sed -i 's/$MYSQL_PASSWORD/'$MYSQL_PASSWORD'/g' /app/config/config.json
-sed -i 's/$MYSQL_HOST/'$MYSQL_HOST'/g' /app/config/config.json
-sed -i 's/$MYSQL_PORT/'$MYSQL_PORT'/g' /app/config/config.json
-sed -i 's/$MYSQL_DATABASE/'$MYSQL_DATABASE'/g' /app/config/config.json
-
+if [ ! -f /app/config ];then
+  mkdir -p /app/config
+  cp /config.json /app/config/config.json
+  sed -i 's/$MYSQL_USER/'$MYSQL_USER'/g' /app/config/config.json
+  sed -i 's/$MYSQL_PASSWORD/'$MYSQL_PASSWORD'/g' /app/config/config.json
+  sed -i 's/$MYSQL_HOST/'$MYSQL_HOST'/g' /app/config/config.json
+  sed -i 's/$MYSQL_PORT/'$MYSQL_PORT'/g' /app/config/config.json
+  sed -i 's/$MYSQL_DATABASE/'$MYSQL_DATABASE'/g' /app/config/config.json
+fi
 set +e
 
 exec "$@"
